# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zero-Wait OPD Kiosk API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Patient {
  id: number;
  name: string;
  age: number;
  gender: string;
  phone: string;
  /** @nullable */
  insuranceId: string | null;
  createdAt: string;
}

export interface PatientInput {
  name: string;
  age: number;
  gender: string;
  phone: string;
  insuranceId?: string;
}

export interface Visit {
  id: number;
  patientId: number;
  symptoms: string;
  suggestedDepartment: string;
  insuranceStatus: string;
  /** @nullable */
  chatSummary: string | null;
  createdAt: string;
}

export interface VisitInput {
  patientId: number;
  symptoms: string;
  suggestedDepartment: string;
  insuranceStatus: string;
  chatSummary?: string;
}

export interface VisitWithPatient {
  id: number;
  patientId: number;
  symptoms: string;
  suggestedDepartment: string;
  insuranceStatus: string;
  /** @nullable */
  chatSummary: string | null;
  createdAt: string;
  patient: Patient;
}

export interface IdTextInput {
  idText: string;
}

export interface ExtractedPatient {
  name: string;
  /** @nullable */
  age: number | null;
  /** @nullable */
  gender: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  insuranceId: string | null;
}

export interface InsuranceCheckInput {
  insuranceId: string;
}

export interface InsuranceCheckResult {
  insuranceId: string;
  status: string;
  details: string;
}

export interface DepartmentCount {
  department: string;
  count: number;
}

export interface KioskStats {
  totalPatients: number;
  visitsToday: number;
  totalVisits: number;
  departmentBreakdown: DepartmentCount[];
  recentVisits: VisitWithPatient[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type SearchPatientsParams = {
q: string;
};