# lib/api-zod/src/generated/types/extractedPatient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zero-Wait OPD Kiosk API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedPatient {
  name: string;
  /** @nullable */
  age: number | null;
  /** @nullable */
  gender: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  insuranceId: string | null;
}