# lib/api-zod/src/generated/types/insuranceCheckResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zero-Wait OPD Kiosk API
 * OpenAPI spec version: 0.1.0
 */

export interface InsuranceCheckResult {
  insuranceId: string;
  status: string;
  details: string;
}