# lib/api-zod/src/generated/types/kioskStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zero-Wait OPD Kiosk API
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentCount } from './departmentCount';
import type { VisitWithPatient } from './visitWithPatient';

export interface KioskStats {
  totalPatients: number;
  visitsToday: number;
  totalVisits: number;
  departmentBreakdown: DepartmentCount[];
  recentVisits: VisitWithPatient[];
}