# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zero-Wait OPD Kiosk API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  age: number;
  gender: string;
  phone: string;
  /** @nullable */
  insuranceId: string | null;
  createdAt: Date;
}